+++ MSAL/src/native_auth/network/errors/MSALNativeAuthErrorMessage.swift
@@ -39,6 +39,8 @@ enum MSALNativeAuthErrorMessage {
     static let invalidUsername = "Invalid username"
     static let generalError = "General error"
     static let invalidCode = "Invalid code"
+    static let notImplemented = "This flow is not implemented yet"
+    static let invalidContinuationToken = "Invalid continuation token"
     static let invalidChallenge = "Invalid challenge"
     static let invalidInput = "Invalid input"
     static let refreshTokenExpired = "Refresh token is expired"